Introduce Raft based IAtomicLong & Raft improvements
Raft improvements are:

- Introduce snapshot into RaftLog
- Move Raft config params to RaftConfig
- Implement Raft pre-vote phase
- Implement membership change logic in the core Raft module
- Add config option to enable `appendNopEntryOnLeaderElection`
- Add local read/query mechanism without appending Raft log

+++ hazelcast-raft/src/main/java/com/hazelcast/raft/impl/dto/AppendFailureResponse.java
@@ -8,16 +8,25 @@ import com.hazelcast.raft.impl.RaftEndpoint;
 
 import java.io.IOException;
 
+/**
+ * Struct for failure response to AppendEntries RPC.
+ * <p>
+ * See <i>5.3 Log replication</i> section of <i>In Search of an Understandable Consensus Algorithm</i>
+ * paper by <i>Diego Ongaro</i> and <i>John Ousterhout</i>.
+ *
+ * @see AppendRequest
+ * @see AppendSuccessResponse
+ */
 public class AppendFailureResponse implements IdentifiedDataSerializable {
 
     private RaftEndpoint follower;
     private int term;
-    private int expectedNextIndex;
+    private long expectedNextIndex;
 
     public AppendFailureResponse() {
     }
 
-    public AppendFailureResponse(RaftEndpoint follower, int term, int expectedNextIndex) {
+    public AppendFailureResponse(RaftEndpoint follower, int term, long expectedNextIndex) {
         this.follower = follower;
         this.term = term;
         this.expectedNextIndex = expectedNextIndex;
@@ -31,7 +40,7 @@ public class AppendFailureResponse implements IdentifiedDataSerializable {
         return term;
     }
 
-    public int expectedNextIndex() {
+    public long expectedNextIndex() {
         return expectedNextIndex;
     }
 
@@ -49,14 +58,14 @@ public class AppendFailureResponse implements IdentifiedDataSerializable {
     public void writeData(ObjectDataOutput out) throws IOException {
         out.writeInt(term);
         out.writeObject(follower);
-        out.writeInt(expectedNextIndex);
+        out.writeLong(expectedNextIndex);
     }
 
     @Override
     public void readData(ObjectDataInput in) throws IOException {
         term = in.readInt();
         follower = in.readObject();
-        expectedNextIndex = in.readInt();
+        expectedNextIndex = in.readLong();
     }
 
     @Override
